pyproject.toml
[tool.poetry]
name = "nonebot_plugin_rollpig"
version = "0.1.7"
description = "每日抽取属于自己的小猪"
authors = ["Bear_lele <lele2587342146@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/Bearlele/nonebot-plugin-rollpig"

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.4.0"
nonebot-adapter-onebot = "^2.2.0"
nonebot-plugin-htmlrender = "^0.4.0"
nonebot-plugin-apscheduler = ">=0.5.0"
nonebot-plugin-localstore = ">=0.7.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.nonebot.adapters]
nonebot-adapter-onebot = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
"@local" = []